+++ src/uk/co/rubendougall/tabletennis/Controller.java
@@ -1,0 +1,4 @@
+package uk.co.rubendougall.tabletennis;
+
+public class Controller {
+}
